Potential fix for code scanning alert no. #281: Workflow does not contain permissions
To fix the issue, we will add a permissions block at the workflow level to define the least privileges required. Based on the tasks performed in the workflow:

The contents: read permission is sufficient for most steps, such as checking out the repository and running Maven commands.
The Test job uses the GITHUB_TOKEN for reporting coverage to Coveralls, which might require contents: read and possibly statuses: write for updating commit statuses.
We will add the following permissions block at the root of the workflow:
```
permissions:
  contents: read
  statuses: write
```
This ensures that the workflow adheres to the principle of least privilege.